// src/types/property.ts
export type PropertyData = {
  code: string;
  totalRoom: number;
  roomRev: number;
  fnbRev: number;
  otherRev: number;
  totalRev: number;
  occPer: number;
};

export type PropertyJSON = {
  property: string;
  totalRoom: string;
  roomRevenue: number;
  fnbRevenue: number;
  otherRevenue: number;
  totalRevenue: number;
  OccPer: number;
};